Move code dealing with persistence to UnreadWidgetRepository
