+++ k9mail/src/main/java/com/fsck/k9/widget/unread/KoinModule.kt
@@ -1,0 +1,7 @@
+package com.fsck.k9.widget.unread
+
+import org.koin.dsl.module.applicationContext
+
+val unreadWidgetModule = applicationContext {
+    bean { UnreadWidgetRepository(get()) }
+}
